Starting off the repository.
Minor cleanups on file encoding and the addition of the license were
added.  No guarantee this actually builds, but it's a start.

Newtonsoft.Json.dll, a dependency, is not included; it's an open
source library and should not be difficult to properly integrate into
the build system.

(This commit and message written on behalf of the author.)